chore(core): move comic-downloader to core
