+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## Draft: 2.3.0
+
+`2023-04-18`
+
+### Deprecation
+
+- Import ComicDownloader from 'comic-downloader.ts' is deprecated, and will be removed in the next major version.
+
+```typescript
+// Your module
+
+// before
+import ComicDownloader from '../../comic-downloader';
+
+// now
+import ComicDownloader from '../../core';
+```
+
+### Feature
+
+- [CLI] New `-S, --shorthand-url` flag is added, please refer to the documentation of each module.
+  - When using the `-h, --history` flag, the url registered will still be the completed url.
+  - A `static urlCompletion` method is added to modules for developers.
+  - The `-u, --url` flag has a higher priorty than this flag.
+- [CLI] New `-l, --list` flag is added, that can download a list of urls.
+  - Please note that due to the stability of each module, you may end up with multiple failures.
+  - You can use the `-h, --history` flag added in the `version 2.2.0` to register your download history, so that you can use this flag as a "sync" functionality.
+  - You can use the `-S, --shorthand-url` flag without value if the list your list is composed of shorthand urls.
+
+### Dev
+
+- The `comic-downloader.ts` is moved to `core/index.ts`, causing the deprecation, but `comic-downloader.ts` will still exist for a while until the next major version.
+- A `static urlCompletion` method is added to modules in order to parse the shorthand urls.
+
 ## 2.2.0
 
 `2023-04-08`
