+++ src/cli/templates.ts
@@ -15,7 +15,7 @@ export function indexDts(name: string) {
 export function indexTs(name: string) {
   return `/// <reference path="./index.d.ts" />
 
-  import ComicDownloader from "../../comic-downloader";
+  import ComicDownloader from "../../core";
   
   export default class ${cap1(name)}Downloader extends ComicDownloader {
     static readonly siteName = "${name}";
