+++ src/index.ts
@@ -3,6 +3,6 @@
  * Copyright (c) 2023 Yan
  */
 
-export * from "./modules";
+export * from './modules';
 
-export { default as default } from "./comic-downloader";
+export { default as default } from './core';
